Move PushUpdateAddDeleteTags to repository module from models
